Remove pandas (#64)
Fixes #34

This removes pandas as a dependency. We were only using it to make
simple sql queries even simpler; However, it caused some error that
was not recreated in tests, and avoiding the problematic version
of pandas limited users to numpy<2.0. This change frees users up to
move to numpy 2.0 without worrying about pandas compatibility.

Along the way, continue to move database access to the `_db` module.